:truck: rename charge to battery-charge and moving to map dir
