+++ src/components/map/battery-charge.scss
@@ -1,4 +1,4 @@
-@import "../resources/scss/variables";
+@import "../../resources/scss/variables";
 
 .zap-icon {
     height: 32px;
